Move AuthorizationAdvisorProxyFactory
To prevent package tangles

Issue gh-14596
